requirements.txt
alembic==1.13.2
fastapi==0.112.1
httpx==0.27.2
mysql-connector-python==9.0.0
pydantic[email]==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
uvicorn==0.30.6
pytest==8.3.2